requirements.txt
Django>=3.2,<4.0
django-cotton>=0.9.0
shadcn-django>=0.1.0
django-tailwind>=3.5
django-debug-toolbar>=3.2
python-decouple>=3.8
